reorganize project structure (entry model)
The project was reorganized to allow for a better
maintainability and clarity. It will follow a
better structure using models and controllers.

Helper functions are outsourced to a utils module
to have them all in one place.

This specific commit deals with the way the JSON
file for the response is generated.