extract functions for building the landing page
Another iteration to enhance the maintainability
and clarity of the web app.

index.html is cleaned and is now consistently build
by the js. No direct input by the Flask app.
Additionally, the deprecated Markup functionality
of flask is not used anymore.

For less confusion the file documentation.html is
now renamed to howto.html. This describes the
purpose of this text better.