Redesign how web page is build
Before this commit, the javascript accessed
5 different endpoints just to get the
necessary information to build the UI. Also,
there was a lot of code repetition in the
javascript.

The app.py is much more concise with a single
endpoint that delivers all data to build the
web page/UI. The same is true for the js.