MOOChub Tool V2
There was no good way to implement the API support
for CoKoMo in the old version. Therefore, the tool
was set up new and only parts of the old structure
were kept.

It is now easier to switch between different
configs.

The new structure especially within the javascript
(js) folder is now more oriented on the usage
instead of controllers and models.